Add separate folders for the individual pages as they will house multiple files relating to the pages (such as tests and page specific components).
